requirements.txt
pytorch-lightning==1.9.3
tqdm
bcolz-zipline
prettytable
menpo
mxnet
numpy==1.23.4
opencv-python
hydra-core
rich
pandas
scikit-image
scikit-learn
wandb
scipy==1.8.1
torch==1.13.1
torchvision==0.14.1

